Extract more properties into IBattle interface
